requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.7.0
pillow>=10.4.0
numpy>=1.26.4,<2.0
firebase-admin==6.5.0
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0
